Consolidate all format-related functions into formatUtil.ts
Related to #1230
