+++ src/app/utils/formatters/index.ts
@@ -1,12 +1,7 @@
-import { formatBalance } from "./formatBalance";
 import { displayFormattedDateWithTimeZone, getNextPaymentDate, getTimeZoneSymbol } from "./dateUtils";
-import { formatChainName, formatAddress } from "./chainUtils";
 
 export {
-  formatBalance,
   displayFormattedDateWithTimeZone,
   getNextPaymentDate,
   getTimeZoneSymbol,
-  formatChainName,
-  formatAddress,
 }
